create header as a shared component
