+++ packages/web/src/components/organisms/PropertyHeader/index.tsx
@@ -1,46 +1,40 @@
 import React from 'react'
-import { Button } from 'antd'
-import { useConnectWallet } from 'src/fixtures/wallet/hooks'
-import { BrandLogo } from 'src/components/atoms/BrandLogo'
+import { Header } from 'src/components/molecules/Header'
 
 interface Props {
   propertyAddress: string
 }
 
 export const PropertyHeader = ({ propertyAddress }: Props) => {
-  const { isConnected, connect } = useConnectWallet()
-  const handleClick = () => {
-    connect()
-  }
-
   return (
     <div
       style={{
         width: 'auto',
         position: 'relative',
         height: '562px',
-        padding: '50px',
         background: `
           url('//raw.githubusercontent.com/dev-protocol/assets/master/property/${propertyAddress}/header.jpg'),
           linear-gradient(111.32deg, #2F80ED 0%, #D5E6FB 100%)`,
         backgroundSize: 'cover'
       }}
     >
-      <BrandLogo colorSchema={'white'} props={{ width: '14em', height: undefined }}></BrandLogo>
-      <Button size="large" style={{ float: 'right' }} disabled={isConnected} onClick={handleClick}>
-        {isConnected && 'Wallet connected'}
-        {!isConnected && 'Connect to a wallet'}
-      </Button>
+      <Header colorSchema={'white'} />
       <div
         style={{
-          position: 'absolute',
-          bottom: '0',
-          padding: '0 0 122px 333px',
-          display: 'block'
+          padding: '50px'
         }}
       >
-        <span style={{ background: 'white', padding: '0.5em' }}>Property Address</span>
-        <div style={{ background: 'white' }}>{propertyAddress}</div>
+        <div
+          style={{
+            position: 'absolute',
+            bottom: '0',
+            padding: '0 0 122px 333px',
+            display: 'block'
+          }}
+        >
+          <span style={{ background: 'white', padding: '0.5em' }}>Property Address</span>
+          <div style={{ background: 'white' }}>{propertyAddress}</div>
+        </div>
       </div>
     </div>
   )
